Promote dynamic_size_type to upper namespace
This removes the dynamic_size_type from the detail namespace, and
promotes it to be in the bit::memory namespace instead for better
visibility
