+++ include/bit/memory/block_allocators/malloc_block_allocator.hpp
@@ -12,7 +12,7 @@
 #include "detail/cached_block_allocator.hpp" // detail::cached_block_allocator
 #include "detail/named_block_allocator.hpp"  // detail::named_block_allocator
 
-#include "../utilities/detail/dynamic_size_type.hpp" // dynamic_size, etc
+#include "../utilities/dynamic_size_type.hpp" // dynamic_size, etc
 #include "../utilities/allocator_info.hpp"    // allocator_info
 #include "../utilities/macros.hpp"            // BIT_MEMORY_UNLIKELY
 #include "../utilities/memory_block.hpp"      // memory_block
@@ -27,7 +27,7 @@ namespace bit {
     namespace detail {
       template<std::size_t Size>
       struct malloc_block_allocator_base
-        : detail::dynamic_size_type<0,Size>
+        : dynamic_size_type<0,Size>
       {
         malloc_block_allocator_base() noexcept = default;
         malloc_block_allocator_base( malloc_block_allocator_base&& ) noexcept = default;
@@ -38,10 +38,10 @@ namespace bit {
 
       template<>
       struct malloc_block_allocator_base<dynamic_size>
-        : detail::dynamic_size_type<0,dynamic_size>
+        : dynamic_size_type<0,dynamic_size>
       {
         explicit malloc_block_allocator_base( std::size_t size )
-          : detail::dynamic_size_type<0,dynamic_size>(size)
+          : dynamic_size_type<0,dynamic_size>(size)
         {
 
         }
@@ -64,7 +64,7 @@ namespace bit {
     class malloc_block_allocator
       : private detail::malloc_block_allocator_base<Size>
     {
-      using block_size_member = detail::dynamic_size_type<0,Size>;
+      using block_size_member = dynamic_size_type<0,Size>;
       using base_type = detail::malloc_block_allocator_base<Size>;
 
       //----------------------------------------------------------------------
